==> R/02_01_exploratory_data_analyses.R <==
# ---------------------------------------------------------------------------------- #
#####              Function for all exploratory data analyses (EDA)              #####
# ---------------------------------------------------------------------------------- #

### Script preparations ________________________________________________________
# I first need to source some of my previous R scripts to be able to call their functions:
source(file = here::here("R/01_02_utility_functions.R"))
source(file = here::here("R/01_03_preparing_data.R"))

library(magrittr) # The only library that I truly need to load (in order to be able to use pipes).

utils::globalVariables("where") # This is necessary for now as tidyselect::where is not an
# exported function!


# IMPORTANT NOTE: because of changes related to successive submissions, the names of the variables
# may vary slightly between the manuscript and the code. For instance, that is the case for the two
# "variables of interest", namely "greenspace area" (called "patch_area" in the code) and
# "greenspace connectivity" (called "log_F_metric_d2b1" or simply "F_metric" in the code). But
# mostly, the names are pretty straightforward to understand.





### _______________________________________________________
#' Exploratory DATA Analyses and final tits data formatting
#'
#' @description (THEORETICALLY MEANT FOR INTERNAL USE ONLY, use with caution!).
#' The `tfinal_EDAta` function loads and modifies the dataset generated by the last praparing
#' function: (\code{\link[tdata_upD_final]{tdata_upD_final}}). It completes these data with
#' a few additional inputs and corrections, and then use them to explore the data's main
#' statistical properties.
#' Finally, the function exports the updated dataset as a tibble or as a new .csv file.
#'
#' @return A list containing nine elements:
#' 1) the updated "final" dataset (a tibble accessible with `tfinal_EDAta()$final_dataset`);
#' 2) the correlation circle from the normed-PCA used to create the "urban intensity" proxy
#' variable (a plot accessible with `tfinal_EDAta()$urban_corplot`);
#' 3) the correlation circle from the normed-PCA used to create the "light pollution" proxy
#' variable (a plot accessible with `tfinal_EDAta()$lightpol_corplot`);
#' 4) the correlation matrix for all the potential predictors (a plot accessible with
#' `tfinal_EDAta()$x_cormat`);
#' 5) the correlation matrix for all the potential response variables (a plot accessible with
#' `tfinal_EDAta()$y_cormat`);
#' 6) the pairplot for all the potential predictors (a plot accessible with
#' `tfinal_EDAta()$x_pairplot`);
#' 7) the pairplot for all the potential response variables (a plot accessible with
#' `tfinal_EDAta()$y_pairplot`);
#' 8) the table showing the skewness and kurtosis for all the potential predictors (accessible
#' with `tfinal_EDAta()$x_skewtable`);
#' 9) the table showing the skewness and kurtosis for all the potential response variables
#' (accessible with `tfinal_EDAta()$y_skewtable`);
#' 10) the path to the exported .csv file of the updated "final" dataset
#' (`ndata_final.csv`; accessible with `tdata_upD_final()$path`).
#'
#' @export
#' @importFrom here here
#' @importFrom readr read_csv2
#' @importFrom readr cols
#' @importFrom readr col_factor
#' @importFrom readr write_csv2
#' @importFrom FactoMineR PCA
#' @importFrom factoextra fviz_pca_var
#' @importFrom gridExtra grid.arrange
#' @importFrom moments skewness
#' @importFrom knitr kable
#' @importFrom stats cor
#' @importFrom stats na.omit
#' @importFrom stats relevel
#' @importFrom stats median
#' @importFrom ggcorrplot cor_pmat
#' @importFrom ggcorrplot ggcorrplot
#' @importFrom ggplot2 theme_gray
#' @importFrom GGally ggpairs
#' @importFrom moments skewness
#' @importFrom moments kurtosis
#' @importFrom lubridate yday
#' @importFrom dplyr mutate
#' @importFrom dplyr filter
#' @importFrom dplyr select
#' @importFrom dplyr inner_join
#' @importFrom dplyr case_when
#' @importFrom dplyr relocate
#' @importFrom dplyr rename
#' @importFrom dplyr across
#' @importFrom tibble as_tibble
#'
#' @examples
#' \dontrun{
#' # To obtain the final dataset and one of the skewness/kurtosis table (the principle is the
#' # same for every other outputs):
#' final_titdata <- tfinal_EDAta()$final_dataset
#' pred_sk.table <- tfinal_EDAta()$x_skewtable
#' }
tfinal_EDAta <- function(){

  ###################### ********************************************* ###########################
  # -------------------------------------- #
  ##### 1. Data import and preparation #####
  # -------------------------------------- #

  ##### * 1.1. Data import and formatting ----------------------------------------
  # ---------------------------------------------------------------------------- #

  .pardefault <- par()

  ### ** 1.1.1. Import ----
  # _______________________

  tits_clean <- tdata_upD_final()$clean_dataset
  tits_clean %>% dplyr::filter(dist == 50) -> tits_clean_50 # Only selects the 50m-based dataset!
  tits_clean %>% dplyr::filter(dist == 100) -> tits_clean_100 # Only selects the 100m-based dataset!
  tits_clean %>% dplyr::filter(dist == 150) -> tits_clean_150 # Only selects the 150m-based dataset!
  tits_clean %>% dplyr::filter(dist == 200) -> tits_clean_200 # Only selects the 200m-based dataset!


  ## IMPORTANT NOTE: The code below only shows the treatments for the "tits_clean_150" dataset
  # (if you want to apply it to the other datasets, you have to manually change the dataset name
  # manually):
  manag <- readr::read_csv2(here::here("data", "veg_manag_factor.csv"), # To import another (late
                            # coming) independent variable.
                            col_names = TRUE, na = "NA",
                            col_types = readr::cols(id_nestbox = readr::col_factor(),
                                                    manag_intensity = readr::col_factor(
                                                      levels = c("0", "1", "2"),
                                                      include_na = FALSE)))
  tits_clean_150 <- dplyr::inner_join(tits_clean_150, manag, by = "id_nestbox")

  ### To import the connectivity metrics computed for both species (using the Bash script provided
  # in the '02_00_graphab_analyses.R' file):
  cmetrics_pm <- readr::read_csv2(here::here("data", "cmetrics_pm.csv"),
                                  col_names = TRUE, na = "NA",
                                  col_types = readr::cols(id_nestbox = readr::col_factor(),
                                                          id_patch = readr::col_factor()))
  cmetrics_pm %>% dplyr::select(-cost_to_patch, -id_patch, -patch_area, -perim,
                                -pmRr_d1_capa2, -pmRr_d2_capa2, -pmRr_d3_capa2) %>%
    dplyr::inner_join(tits_clean_150, cmetrics_pm, by = "id_nestbox", multiple = "all") -> ttt # Only
  # selects some metrics!

  cmetrics_cc <- readr::read_csv2(here::here("data", "cmetrics_cc.csv"),
                                  col_names = TRUE, na = "NA",
                                  col_types = readr::cols(id_nestbox = readr::col_factor(),
                                                          id_patch = readr::col_factor()))
  cmetrics_cc %>% dplyr::select(-cost_to_patch, -ccRr_d1_capa2, -ccRr_d2_capa2, -ccRr_d3_capa2) %>%
    dplyr::inner_join(ttt, cmetrics_cc, by = "id_nestbox", multiple = "all") -> ttt # Only selects
  # some metrics!


  ### ** 1.1.2. Binding ----
  # ________________________

  tits_clean_150$age_num <- as.numeric(as.character(tits_clean_150$age_class))
  tits_clean_150$strata_num <- as.numeric(as.character(tits_clean_150$strata_div))
  tits_clean_150$manag_num <- as.numeric(as.character(tits_clean_150$manag_intensity))
  tits_clean_150 %>%  dplyr::relocate(manag_intensity, .after = age_class) -> tits_clean_150
  tits_clean_150 %>%  dplyr::relocate(manag_num, .after = manag_intensity) -> tits_clean_150
  tits_clean_150 %>%  dplyr::relocate(age_num, .after = age_class) %>%
    dplyr::rename(traffic = trafic) -> tits_clean_150


  ## IMPORTANT NOTE: here also, I only show the code for the "tits_clean_150" dataset:
  ntits200 <- cbind(tits_clean_200, ttt[,2:26]) # To select all patch related variables!
  ntits200 %>% dplyr::select(-dist, -breeding_window, -father_id, -mother_id, -mean_winter_t,
                          -sd_winter_t, -lsource_vs150_m, -lsource_vs150_iq,
                          -soft_manag_area) -> ntits200
  rm(cmetrics_pm, cmetrics_cc, manag, tits_clean)
  ntits200 <- tibble::as_tibble(ntits200)

  # Weighting of woodyveg_volume with strata_num:
  ntits200 %>% dplyr::mutate(strata_w = dplyr::case_when(
    strata_div == "0" ~ 0.25,
    strata_div == "1" ~ 0.5,
    strata_div == "2" ~ 0.6,
    strata_div == "3" ~ 0.75,
    strata_div == "4" ~ 0.9)) %>%
    dplyr::mutate(woodyveg_vw = woodyveg_volume*strata_w) %>%
    dplyr::select(-strata_div, -strata_num, -strata_w) %>%
    dplyr::relocate(woodyveg_vw, .after = woodyveg_sd) %>%
    dplyr::rename(patch_perim = perim) -> ntits200



  ##### * 1.2. Data reduction ----------------------------------------------------
  # ---------------------------------------------------------------------------- #

  ### For variables related to vegetation characteristics_________________________#
  # Does not yield meaningful results.

  ### For variables related to landscape composition______________________________#
  ntits200 %>% dplyr::select(id_nestbox, site, built_area, build_volume,
                          build_sd, open_area) -> xxx # In the end, I decided to remove traffic!

  # Normed-PCA:
  res.pca <- FactoMineR::PCA(X = xxx[, 3:ncol(xxx)], scale.unit = TRUE, graph = FALSE)
  # To plot results:
  landscape.varplot <- factoextra::fviz_pca_var(res.pca, col.var = "contrib",
                                                gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07"),
                                                repel = TRUE)
  landscape.indplot <- plot(res.pca, choix = "ind", autoLab = "yes")
  #gridExtra::grid.arrange(landscape.varplot, landscape.indplot, ncol = 2)

  # As the first axis (PC) of my PCA satisfactorily synthesizes a large amount of the
  # variance (70.2%)*** of my four variables, we can use the coordinates of observations on
  # this axis as a synthetic variable:
  zzz <- res.pca$ind$coord[,1]
  ntits200$urban_intensity <- zzz # This variable opposes nestboxes located in very dense urban
  # areas with less densely urbanised (or even not-urbanised) areas.



  ### For variables related to light pollution____________________________________#
  ntits200 %>% dplyr::select(id_nestbox, site,
                          lsource_0_m, lsource_10_m, lflux_0_m, lflux_10_m) -> xxx

  # Normed-PCA:
  res.pca <- FactoMineR::PCA(X = xxx[, 3:ncol(xxx)], scale.unit = TRUE, graph = FALSE)
  # To plot results:
  lightpol.varplot <- factoextra::fviz_pca_var(res.pca, col.var = "contrib",
                                               gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07"),
                                               repel = TRUE)
  lightpol.indplot <- plot(res.pca, choix = "ind", autoLab = "yes")
  #gridExtra::grid.arrange(lightpol.varplot, lightpol.indplot, ncol = 2)

  # As the first axis (PC) of my PCA satisfactorily synthesizes a large amount of the variance
  # (75.7%)*** of my four variables, we can use the coordinates of observations on this axis as
  # a synthetic variable:
  zzz <- res.pca$ind$coord[,1]
  ntits200$light_pollution <- zzz # This variable opposes nestboxes located in areas experiencing
  # strong light pollution against more preserved areas.



  ##### * 1.3. Datasets formatting -----------------------------------------------
  # ---------------------------------------------------------------------------- #

  ntits200[which(ntits200$noise_m == min(ntits200$noise_m)), "noise_m"] <- 31
  ntits200 %>% dplyr::mutate(
    F_metric_d1b0 = dplyr::case_when(
      species == "PM" ~ pmF_d1_beta0, species == "CC" ~ ccF_d1_beta0),
    F_metric_d2b0 = dplyr::case_when(
      species == "PM" ~ pmF_d2_beta0, species == "CC" ~ ccF_d2_beta0),
    F_metric_d3b0 = dplyr::case_when(
      species == "PM" ~ pmF_d3_beta0, species == "CC" ~ ccF_d3_beta0),
    F_metric_d1b1 = dplyr::case_when(
      species == "PM" ~ pmF_d1_beta1, species == "CC" ~ ccF_d1_beta1),
    F_metric_d2b1 = dplyr::case_when(
      species == "PM" ~ pmF_d2_beta1, species == "CC" ~ ccF_d2_beta1),
    F_metric_d3b1 = dplyr::case_when(
      species == "PM" ~ pmF_d3_beta1, species == "CC" ~ ccF_d3_beta1),
    Rr_metric_d1c1 = dplyr::case_when(
      species == "PM" ~ pmRr_d1_capa1, species == "CC" ~ ccRr_d1_capa1),
    Rr_metric_d2c1 = dplyr::case_when(
      species == "PM" ~ pmRr_d2_capa1, species == "CC" ~ ccRr_d2_capa1),
    Rr_metric_d3c1 = dplyr::case_when(
      species == "PM" ~ pmRr_d3_capa1, species == "CC" ~ ccRr_d3_capa1),
    Dr_metric_c1 = dplyr::case_when(
      species == "PM" ~ pmDr_capa1, species == "CC" ~ ccDr_capa1),
    Dr_metric_c2 = dplyr::case_when(
      species == "PM" ~ pmDr_capa2, species == "CC" ~ ccDr_capa2),
  ) %>%
    dplyr::select(id_nestbox, id_patch, site, coord_x, coord_y, year, species, laying_date,
                  flight_date, clutch_size, brood_size, fledgling_nb, mass, tarsus_length,
                  wing_length, F_metric_d1b0, F_metric_d2b0, F_metric_d3b0, F_metric_d1b1,
                  F_metric_d2b1, F_metric_d3b1, Rr_metric_d1c1, Rr_metric_d2c1, Rr_metric_d3c1,
                  Dr_metric_c1, Dr_metric_c2, woodyveg_volume, woodyveg_vw, woodyveg_sd,
                  woody_area, patch_area, patch_perim, herbaceous_area, manag_intensity,
                  light_pollution, noise_m, noise_iq,
                  traffic, built_area, build_volume, build_sd, open_area, water_area, urban_intensity,
                  cumdd_30, cumdd_60, cumdd_between, min_t_before, min_t_between) -> ntits200_reduced

  rm(res.pca, xxx, zzz, ttt)





  ###################### ********************************************* ###########################
  # --------------------------- #
  ##### 2. Data exploration #####
  # --------------------------- #

  ##### * 2.1. Univariate outliers -----------------------------------------------
  # ---------------------------------------------------------------------------- #

  uni.boxplots(ntits150_reduced[,16:ncol(ntits150_reduced)]) # Only for IVs, not Ys.
  uni.dotplots(ntits150_reduced[,16:ncol(ntits150_reduced)])
  # We can see that:
  # - There are extreme values for many variables.
  # - Many variables are quite strongly left-skewed.
  # - As expected, "water_area" has a very strange distribution.
  # - Otherwise, the IVs look relatively nice.



  ##### * 2.2. Normality, skewness and kurtosis ----------------------------------
  # ---------------------------------------------------------------------------- #

  uni.histograms(ntits150_reduced[,16:ncol(ntits150_reduced)])

  ntits150.x <- ntits150_reduced[,16:ncol(ntits150_reduced)]
  ntits150.xnum <- ntits150.x[, sapply(ntits150.x, is.numeric)] # To only select numeric variables.
  tab <- data.frame(moments::skewness(x = ntits150.xnum), moments::kurtosis(x = ntits150.xnum)-3)
  ntits150x_skewkurtable <- knitr::kable(x = tab, digits = 3,
                                      col.names = c("Skewness", "Excess kurtosis"))
  # If skewness values are fairly acceptable for most variables (although F_metric_d3b1
  # and water_area > 3), several variables present (sometimes very) excessive kurtosis:
  # - K > 3: F_metric_d1b0, F_metric_d2b1, Dr_metric_c1, patch_area, build_volume.
  # - K > 4: F_metric_d1b1, woodyveg_volume, noise_m.
  # - K > 5: woodyveg_vw, woodyveg_volume, water_area.
  # - K > 10: F_metric_d3b1 !
  rm(tab)








  ##### * 2.3. Multivariate relationships ----------------------------------------
  # ---------------------------------------------------------------------------- #

  ### ** 2.3.1. For the same buffer distance (same dataset) ----
  # ____________________________________________________________
  ntits150.xnum %>%
    dplyr::select(-F_metric_d3b0, -F_metric_d3b1, -Rr_metric_d1c1, -woodyveg_volume) -> ntits150.xnum
  # To compute the correlation matrix:
  res.cor.ntits150x <- round(stats::cor(ntits150.xnum, use = "complete.obs", method = "spearman"), 2)
  # To compute a matrix of correlation p-values:
  res.pcor.ntits150x <- ggcorrplot::cor_pmat(x = ntits150.xnum, method = "spearman")

  ntits150x.corplot <- ggcorrplot::ggcorrplot(res.cor.ntits150x, type = "upper",
                                           outline.col = "white",
                                           ggtheme = ggplot2::theme_gray,
                                           colors = c("#6D9EC1", "white", "#E46726"),
                                           p.mat = res.pcor.ntits150x,
                                           insig = "blank")
  # We can see that:
  # - All connectivity metrics and woodyveg variables are quite strongly positively correlated,
  #   but negatively with the urban intensity and pollution variables. That's quite unsurprising,
  #   the more woody vegetation in the landscape, the more connectivity, and the less room for
  #   urban features such as buildings, roads, pollution emissions, etc.
  # - Herbaceous areas are less strongly correlated that woody areas.
  # - Patch area and perimeters are very strongly correlated!
  # - Quite logically, the pollution variables are positively correlated with other urban related
  #   ones.
  # - We see that temperature variables are not much correlated with any variable.
  # - Interestingly, "min_t_before" is slightly correlated with most variables when "min_t_between"
  #   is not! It also shows that the minimal temperature before laying tends to be higher when there
  #   is not much vegetation (i.e. when the stations are located in densely urbanised areas).

  ntits150x.pairplot <- GGally::ggpairs(ntits150.xnum)
  # We find again the same patterns. Yet, we can see that:
  # - Some relationships are actually not linear but curvilinear.
  # - There may be problematic outliers, particularly for noise_m.

  ### CONCLUSION: we decided to focus mainly on the "F_metric_d2b1" and "patch_area" as focal
  # proxies of our variables of interest (connectivity and the size of the nesting/foraging
  # habitats) for three reasons: (i) they were the ones that made most sense (also, it's easier
  # to stick with the "F-metric" with a regular weighting (i.e. beta 1) than to explain in a paper
  # that we specifically invented new connectivity metrics (i.e. Rr and Dr) or that we explored
  # fancy parametrisations); (ii) they do not present crippling properties; and (iii) they are
  # among the proxies showing the least correlation with other variables (which could help
  # preventing multicollinearity).



  ### ** 2.3.2. Across buffer distances (across the four datasets) ----
  # ___________________________________________________________________
  ntits50.xnum %>%
    dplyr::select(-F_metric_d3b0, -F_metric_d3b1, -Rr_metric_d1c1, -woodyveg_volume) -> ntits50.xnum
  ntits100.xnum %>%
    dplyr::select(-F_metric_d3b0, -F_metric_d3b1, -Rr_metric_d1c1, -woodyveg_volume) -> ntits100.xnum
  ntits150.xnum %>%
    dplyr::select(-F_metric_d3b0, -F_metric_d3b1, -Rr_metric_d1c1, -woodyveg_volume) -> ntits150.xnum
  ntits200.xnum %>%
    dplyr::select(-F_metric_d3b0, -F_metric_d3b1, -Rr_metric_d1c1, -woodyveg_volume) -> ntits200.xnum

  ntits50.xnum %>%
    dplyr::select(F_metric_d2b1, patch_area, woody_area, herbaceous_area, open_area,
                  water_area, urban_intensity, min_t_between, light_pollution,
                  noise_m, traffic) %>%
    dplyr::rename("flux metric" = F_metric_d2b1,
                  "local patch area" = patch_area,
                  "temperature" = min_t_between,
                  "noise pollution" = noise_m) -> ntits50.xnum_2
  ntits100.xnum %>%
    dplyr::select(F_metric_d2b1, patch_area, woody_area, herbaceous_area, open_area,
                  water_area, urban_intensity, min_t_between, light_pollution,
                  noise_m, traffic) %>%
    dplyr::rename("flux metric" = F_metric_d2b1,
                  "local patch area" = patch_area,
                  "temperature" = min_t_between,
                  "noise pollution" = noise_m) -> ntits100.xnum_2
  ntits150.xnum %>%
    dplyr::select(F_metric_d2b1, patch_area, woody_area, herbaceous_area, open_area,
                  water_area, urban_intensity, min_t_between, light_pollution,
                  noise_m, traffic) %>%
    dplyr::rename("flux metric" = F_metric_d2b1,
                  "local patch area" = patch_area,
                  "temperature" = min_t_between,
                  "noise pollution" = noise_m) -> ntits150.xnum_2
  ntits200.xnum %>%
    dplyr::select(F_metric_d2b1, patch_area, woody_area, herbaceous_area, open_area,
                  water_area, urban_intensity, min_t_between, light_pollution,
                  noise_m, traffic) %>%
    dplyr::rename("flux metric" = F_metric_d2b1,
                  "local patch area" = patch_area,
                  "temperature" = min_t_between,
                  "noise pollution" = noise_m) -> ntits200.xnum_2
  colnames(ntits50.xnum_2) <- paste(colnames(ntits50.xnum_2), "50m", sep = "_")
  colnames(ntits100.xnum_2) <- paste(colnames(ntits100.xnum_2), "100m", sep = "_")
  colnames(ntits150.xnum_2) <- paste(colnames(ntits150.xnum_2), "150m", sep = "_")
  colnames(ntits200.xnum_2) <- paste(colnames(ntits200.xnum_2), "200m", sep = "_")
    ntits_all <- cbind(ntits50.xnum_2, ntits100.xnum_2, ntits150.xnum_2, ntits200.xnum_2)


  # To compute the correlation matrix:
  res.cor.ntitsallx <- round(stats::cor(ntits_all, use = "complete.obs", method = "spearman"), 2)
  # To compute a matrix of correlation p-values:
  res.pcor.ntitsallx <- ggcorrplot::cor_pmat(x = ntits_all, method = "spearman")

  ntits150x.corplot <- ggcorrplot::ggcorrplot(res.cor.ntitsallx, type = "upper",
                                              outline.col = "white",
                                              ggtheme = ggplot2::theme_gray,
                                              colors = c("#6D9EC1", "white", "#E46726"),
                                              p.mat = res.pcor.ntitsallx,
                                              insig = "blank")




  ###################### ********************************************* ###########################
  # ---------------------------------------------------------- #
  ##### 3. Exploration of the potential response variables #####
  # ---------------------------------------------------------- #

  # In this section, we simply explore some properties of our "potential" response variables
  # to see whether something is wrong with them. Of course, we explored them without looking
  # at their relationships with the predictors, otherwise that would have been "cherry-picking".



  ##### * 3.1. Outliers and distributions ----------------------------------------
  # ---------------------------------------------------------------------------- #

  ntits150.y <- ntits150_reduced[,10:15]

  ### For NTITS___________________________________________________________________#

  uni.boxplots(ntits150.y)
  uni.dotplots(ntits150.y)
  # we can say that the 6 DV (dependant variables) have relatively nice distributions although
  # we seem to have a zero-inflation for "brood_size" and "fledgling_nb", as well as a slight
  # left-skewness for the morphometric variables (possibly linked to the differences among tit
  #  species).
  tab <- data.frame(moments::skewness(x = stats::na.omit(ntits150.y)),
                    moments::kurtosis(x = stats::na.omit(ntits150.y))-3)
  ntits150y_skewkurtable <- knitr::kable(x = tab, digits = 3, col.names = c("Skewness",
                                                                         "Excess kurtosis"))
  # But skewness and kurtosis values are very satisfactory.



  ##### * 3.2. Multivariate relationships ----------------------------------------
  # ---------------------------------------------------------------------------- #

  # To compute the correlation matrix:
  res.cor.ntits150y <- round(stats::cor(ntits150.y, use = "complete.obs", method = "spearman"), 2)
  # To compute a matrix of correlation p-values:
  res.pcor.ntits150y <- ggcorrplot::cor_pmat(x = ntits150.y, method = "spearman")

  ntits150y.corplot <- ggcorrplot::ggcorrplot(res.cor.ntits150y, type = "upper",
                                           outline.col = "white",
                                           ggtheme = ggplot2::theme_gray,
                                           colors = c("#6D9EC1", "white", "#E46726"),
                                           p.mat = res.pcor.ntits150y,
                                           insig = "blank")
  # We can see that:
  # - The 3 reproductive variables are strongly positively correlated and so are the three
  #   morphometric ones!
  # - Both "clutch_size" and "brood_size" seem to be negatively correlated with the morphometric
  #   variables suggesting that when large clutches or broods lead to smaller chicks!
  # - Interestingly, the number of fledglings seem uncorrelated with nestlings morphology.

  ntits150y.pairplot <- GGally::ggpairs(ntits150.y)
  # We can see that:
  # - The positive relationship among the morphometric variables is undeniable.
  # - The zero-inflation may be due to another process as the total failures in "brood_size"
  #   or "fledgling_nb" (i.e. 0's) appear unrelated with the numbers of the preceding processes
  #   (e.g. egg laying or hatching).
  # - The negative correlation between "clutch_size" and "brood_size" and the morphometric
  #   variables is not so clear here but seem to exist nonetheless. However, it may be an artefact
  #   of "species" because CC is smaller but tends to be more productive (see below).

  # Disclaimer, remind that:
  # - PM: CS ranges 2-12 (mode 8); BS 0-12 (8); FN 0-11 (0 or 6);
  #       MA 9.9-18.4 (mean 14.8); TL 15.7-20.6 (18.9); WL 27-50.4 (41.5).
  # - CC: CS ranges 2-14 (mode 11); BS 0-14 (10); FN 0-12 (0 or 8);
  #       MA 5.8-13 (mean 9); TL 12-17.7 (15.8); WL 17.5-40.1 (32.9).





  ###################### ********************************************* ###########################
  # -------------------------------- #
  ##### 4. Final data formatting #####
  # -------------------------------- #

  rm(tab, ntits150.x, ntits150.y, ntits150.xnum, res.cor.ntits150x, res.pcor.ntits150x)

  ntits150_reduced %>% dplyr::mutate(Dr_metric_c1 = Dr_metric_c1/1000, # Changing scale of the
                                  # cost-distance.
                                  Dr_metric_c2 = Dr_metric_c2/1000, # Changing scale of the
                                  # cost-distance.
                                  noise_m = noise_m/10, # Converting dB into B.
                                  cumdd_30 = cumdd_30/100, # Converting degree-days into hundred
                                  # of degree-days.
                                  cumdd_between = cumdd_between/100,
                                  traffic = traffic/1000000, # Conversion to millions of
                                  # metre.vehicles per day!
                                  built_area = built_area/1000, # Converting m2 into km2.
                                  built_volume = build_volume/1000, # Converting m3 into dm3.
                                  open_area = open_area/1000) %>%
    dplyr::mutate(brood_size = round(brood_size, digits = 0), # There was a decimal count.
                  log_F_metric_d1b0 = log10(F_metric_d1b0), # Predictors rescaling or normalisation.
                  log_F_metric_d2b0 = log10(F_metric_d2b0),
                  log_F_metric_d3b0 = log10(F_metric_d3b0),
                  log_F_metric_d1b1 = log10(F_metric_d1b1),
                  log_F_metric_d2b1 = log10(F_metric_d2b1),
                  log_F_metric_d3b1 = log10(F_metric_d3b1),
                  log_woody_vol = log10(woodyveg_volume),
                  log_woody_vw = log10(woodyveg_vw),
                  log_woody_area = log10(woody_area),
                  log_patch_area = log10(patch_area),
                  log_patch_perim = log10(patch_perim),
                  log_herb_area = log10(herbaceous_area),
                  sqrt_built_vol = sqrt(built_volume)) %>%
    dplyr::mutate(brood_size = as.integer(brood_size),
                  hatching_rate = brood_size/clutch_size,
                  laying_day = lubridate::yday(laying_date)/10) %>% # Computing and rescaling
    # laying day!
    dplyr::mutate(year = stats::relevel(x = year, ref = 3), # Assign 2019 as the reference group.
                  species = stats::relevel(x = species, ref = 2)) %>% # Assign PM as the
    # reference group.
    dplyr::mutate(manag_low = ifelse(manag_intensity == "0", "1", "0"),
                  manag_mid = ifelse(manag_intensity == "1", "1", "0"),
                  manag_high = ifelse(manag_intensity == "2", "1", "0")) %>%
    dplyr::mutate(dplyr::across(where(is.matrix), as.numeric),
                  dplyr::across(where(is.character), as.factor)) %>%
    dplyr::mutate(coord_y = jitter(x = coord_y, factor = 1.2)) %>%
    dplyr::mutate(coord_x = jitter(x = coord_x, factor = 1.2)) -> ntits1502

  # To enable exploring some interaction models, I further median-centre some of the variables:
  ntits1502 %>% dplyr::mutate(c.log_patch_area = log_patch_area-stats::median(log_patch_area),
                           c.log_woody_vol = log_woody_vol-stats::median(log_woody_vol),
                           c.log_woody_area = log_woody_area-stats::median(log_woody_area),
                           c.log_F_metric_d2b1 = log_F_metric_d2b1-stats::median(log_F_metric_d2b1),
                           c.log_F_metric_d2b0 = log_F_metric_d2b0-stats::median(log_F_metric_d2b0),
                           c.clutch_size = clutch_size-stats::median(clutch_size),
                           c.laying_day = laying_day-stats::median(laying_day),
                           c.sqrt_built_vol = sqrt_built_vol-stats::median(sqrt_built_vol),
                           c.built_area = built_area-stats::median(built_area),
                           c.open_area = open_area-stats::median(open_area),
                           c.log_herb_area = log_herb_area-stats::median(log_herb_area),
                           c.noise_m = noise_m-stats::median(noise_m),
                           c.traffic = traffic-stats::median(traffic),
                           c.cumdd_30 = cumdd_30-stats::median(cumdd_30),
                           c.cumdd_between = cumdd_between-stats::median(cumdd_between),
                           c.min_t_between = min_t_between-stats::median(min_t_between)) -> ntits150_2
  ntits150_2[ntits150_2$id_nestbox == "DIJ-205" & ntits150_2$year == "2019", "species"] <- "PM" # Misidentified
  # observation.




  ##### To export the results
  # _________________________

  readr::write_csv2(x = ntits150_2, file = here::here("output", "tables", "ndata_final.csv"))

  output <- list(final_dataset = ntits150_2,
                 urban_corplot = landscape.varplot,
                 lightpol_corplot = lightpol.varplot,
                 x_cormat = ntits150x.corplot,
                 y_cormat = ntits150y.corplot,
                 x_pairplot = ntits150x.pairplot,
                 y_pairplot = ntits150y.pairplot,
                 x_skewtable = ntits150x_skewkurtable,
                 y_skewtable = ntits150y_skewkurtable,
                 path = here::here("output", "tables", "ndata_final.csv"))
  return(output)
}
